move remaining link helpers out of main module
