+++ lib/hdk_graph_helpers/src/link_helpers.rs
@@ -8,23 +8,25 @@
  * @since   2019-07-03
  */
 
+use std::convert::{ TryInto, TryFrom };
 use hdk::{
     holochain_core_types::{
         cas::content::Address,
         json::JsonString,
-        entry::{
-            entry_type::AppEntryType,
-        },
+        entry::AppEntryValue,
+        entry::entry_type::AppEntryType,
+        error::HolochainError,
     },
-    error::{ ZomeApiResult },
+    error::{ ZomeApiError, ZomeApiResult },
+    link_entries,
+    call,
 };
+use holochain_core_types_derive::{ DefaultJson };
 
 use super::{
     records::{
         create_base_entry,
     },
-    link_entries_bidir,
-    link_remote_entries,
 };
 
 /// Toplevel method for triggering a link creation flow between two records in
@@ -170,4 +172,51 @@ pub fn request_remote_query_index(
 
     // :TODO: decode and return status codes from link requests
     Ok(vec![])
+}
+
+/// Creates a bidirectional link between two entry addresses, and returns a vector
+/// of the addresses of the (respectively) forward & reciprocal links created.
+pub fn link_entries_bidir<S: Into<String>>(
+    source: &Address,
+    dest: &Address,
+    link_type: S,
+    link_name: S,
+    link_type_reciprocal: S,
+    link_name_reciprocal: S,
+) -> Vec<Address> {
+    vec! [
+        link_entries(source, dest, link_type, link_name).unwrap(),
+        link_entries(dest, source, link_type_reciprocal, link_name_reciprocal).unwrap(),
+    ]
+}
+
+/// Common request format for linking remote entries in cooperating DNAs
+#[derive(Serialize, Deserialize, Debug, DefaultJson, Clone)]
+struct RemoteEntryLinkRequest {
+    base_entry: Address,
+    target_entries: Vec<Address>,
+}
+
+/// Calls into a neighbouring DNA to link a "base entry" for the given entry ID
+/// to multiple target entries, via a zome API request conforming to `RemoteEntryLinkRequest`.
+/// This enables the DNA holding the target entries to setup data structures
+/// for querying the associated remote entry records back out.
+fn link_remote_entries<R>(
+    target_dna_id: &str,
+    zome_name: &str,
+    cap_token: &Address,
+    fn_name: &str,
+    base_entry: &Address,
+    target_entries: &Vec<Address>,
+) -> ZomeApiResult<R>
+  where R: TryFrom<AppEntryValue>,
+{
+    let result = call(target_dna_id, zome_name, cap_token.to_owned(), fn_name, RemoteEntryLinkRequest {
+        base_entry: base_entry.clone().into(),
+        target_entries: target_entries.clone().into(),
+    }.into())?;
+
+    result.try_into().map_err(|_| {
+        ZomeApiError::Internal("Could not convert link_remote_entries result to requested type".to_string())
+    })
 }
